requirements.txt
Jinja2>=3.1.0
Markdown>=3.5.0
PyYAML>=6.0
